fix(site): wrap cell outputs in scrollable div
When rendering a cell, Quarto creates a div whose children are the cell
code and cell output. Unfortunately, the cell output is represented as
multiple divs. Outputs to stdout and stderr are separate divs. In my
notebook's case, there are a lot of interleaved stdout and stderr
outputs, which means that there are a ton of output divs in the cell div.
These outputs divs are siblings of the cell's code div, which makes the
scrolling experience horrible (each output div has its own scrollbar).

This commit introduces a Lua filter to wrap output divs with a single
div so that the scrollbar is shared among them. Styles are adjusted
accordingly.